+++ orttraining/orttraining/python/training/__init__.py
@@ -5,5 +5,7 @@
 from onnxruntime.capi._pybind_state import TrainingParameters
 from onnxruntime.capi.training.training_session import TrainingSession
 
+
 from .orttrainer_options import ORTTrainerOptions
 from . import model_desc_validation
+from . import amp
